switch postArtists to NumuClient & delete JSONClient
